Move sql/model.go into pkg/model/
